+++ fabric/src/index.ts
@@ -1,4 +1,3 @@
-export * from './components/Avatar'
 export * from './components/Box'
 export * from './components/Button'
 export * from './components/Card'
@@ -24,6 +23,7 @@ export * from './components/StatusChip'
 export * from './components/Tabs'
 export { Text, TextProps } from './components/Text'
 export * from './components/TextInput'
+export * from './components/Thumbnail'
 export * from './components/Toast'
 export * from './components/Tooltip'
 export * from './icon/index'
